Fixed sign warning

+++ src/pde/grid/hypergrid/hypergrid.cc.h
@@ -137,7 +137,7 @@ std::vector<size_t> HyperGrid<T>::toLoc(size_t index) const {
 
 template<class T>
 std::vector<size_t> HyperGrid<T>::toLoc(const T *obj) const {
-    if (obj < data || obj - data >= gridSize)
+    if (obj < data || obj - data >= (long int)gridSize)
         throw std::runtime_error("Bad location conversion, invalid pointer");
     return toLoc(obj - data);
 }
